(bond) style: move to shared theme config
