+++ apps/bond/components/Home/index.jsx
@@ -1,7 +1,8 @@
-import { Paths } from 'components/Paths';
 import styled from 'styled-components';
+
 import { Hero } from 'components/Home/Hero';
-import StyledMain from 'components/GlobalStyles/StyledMain';
+import { Paths } from 'components/Paths';
+import { StyledMain } from 'components/StyledMain';
 
 const PathsContainer = styled.div`
   margin: auto 0;
