+++ apps/bond/components/Layout/index.jsx
@@ -24,7 +24,7 @@ const { Header, Content } = AntdLayout;
 const endpoint = process.env.NEXT_PUBLIC_SOLANA_MAINNET_BETA_URL;
 
 const StyledHeader = styled(Header)`
-  border-bottom: 1px solid ${COLOR.BORDER_GREY};
+  border-bottom: 1px solid ${COLOR.BORDER_GREY_2};
 `;
 
 const Layout = ({ children }) => {
