+++ apps/bond/components/StyledMain.jsx
@@ -1,10 +1,8 @@
 import styled from 'styled-components';
 
-const StyledMain = styled.main`
+export const StyledMain = styled.main`
   display: flex;
   flex-direction: column;
   max-width: 1400px;
   margin: 0 auto;
 `;
-
-export default StyledMain;
